requirements.txt
discord.py
flake8
pytest
coverage
pytest-cov
rope
pep8